Update games.lua
 whether the number is odd or even, or if the numbers are high (19–36) or low (1–18).

To determine the winning number, a croupier spins a wheel in one direction, then spins a ball in the opposite direction around a tilted circular track running around the outer edge of the wheel. The ball eventually loses momentum, passes through an area of deflectors, and falls onto the wheel and into one of 37 (single zero French/European style roulette) or 38 (double zero American style roulette) colored and numbered pockets on the wheel. The winnings are then paid to anyone who has placed a successful bet.